requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
cymysql==0.9.13
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-cymysql==2.1.0
django-filter==2.1.0
django-heroku==0.3.1
django-mysql==2.4.1
django-rest-auth==0.9.3
django-toolbelt==0.0.1
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
oauthlib==3.0.0
psycopg2==2.7.7
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
static3==0.7.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
